refactor: seperate withIosAlternativeBilling
